Implemented the new YAML data store implementation. Ready for testing next session!
Refactored a few classes to look nicer, cleaned up various bits of config-related logic. Gonna need to refactor a bunch more later down the line, but want to get this on the road first to avoid overcomplicating things.

+++ src/main/java/com/sulphate/chatcolor2/data/PlayerData.java
@@ -6,10 +6,10 @@ public class PlayerData {
 
     private final UUID uuid;
     private String colour;
-    private int defaultCode;
+    private long defaultCode;
     private boolean dirty;
 
-    public PlayerData(UUID uuid, String colour, int defaultCode) {
+    public PlayerData(UUID uuid, String colour, long defaultCode) {
         this.uuid = uuid;
         this.colour = colour;
         this.defaultCode = defaultCode;
@@ -27,11 +27,11 @@ public class PlayerData {
         this.colour = colour;
     }
 
-    public int getDefaultCode() {
+    public long getDefaultCode() {
         return defaultCode;
     }
 
-    public void setDefaultCode(int defaultCode) {
+    public void setDefaultCode(long defaultCode) {
         this.defaultCode = defaultCode;
     }
